@@ -2,8 +2,8 @@ namespace PayrollServer.Application.DTOs.Request
 {
     public class ChangePasswordRequest
     {
-        public string CurrentPassword { get; set; }
-        public string NewPassword { get; set; }
-        public string ConfirmPassword { get; set; }
+        public required string CurrentPassword { get; set; } 
+        public required string NewPassword { get; set; }
+        public required string ConfirmPassword { get; set; }
     }
 } 
