@@ -2,8 +2,8 @@ namespace PayrollServer.Application.DTOs.Request
 {
     public class LoginRequest
     {
-        public string Email { get; set; }
-        public string Password { get; set; }
+        public required string Email { get; set; }
+        public required string Password { get; set; }
         public bool RememberMe { get; set; }
     }
 } 
