@@ -4,12 +4,12 @@ namespace PayrollServer.Application.DTOs.Request
 {
     public class RegisterUserRequest
     {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Email { get; set; }
-        public string Username { get; set; }
-        public string Password { get; set; }
-        public string ConfirmPassword { get; set; }
+        public required string FirstName { get; set; }
+        public required string LastName { get; set; }
+        public required string Email { get; set; }
+        public required string Username { get; set; }
+        public required string Password { get; set; }
+        public required string ConfirmPassword { get; set; }
         public DateTime DateOfBirth { get; set; }
     }
 } 
