@@ -5,11 +5,11 @@ namespace PayrollServer.Application.DTOs.Response
 {
     public class UserProfileResponse
     {
-        public string UserId { get; set; }
-        public string Username { get; set; }
-        public string Email { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
+        public required string UserId { get; set; }
+        public required string Username { get; set; }
+        public required string Email { get; set; }
+        public required string FirstName { get; set; }
+        public required string LastName { get; set; }
         public DateTime DateOfBirth { get; set; }
         public DateTime CreatedAt { get; set; }
         public IList<string> Roles { get; set; } = new List<string>();
