@@ -24,67 +24,82 @@ namespace PayrollServer.Infrastructure.Extensions
             this IServiceCollection services,
             IConfiguration configuration)
         {
-            // Configure DbContext
-            var connectionString = configuration.GetConnectionString("DefaultConnection");
-            services.AddDbContext<ApplicationDbContext>(options =>
-                options.UseSqlServer(
-                    connectionString,
-                    b => b.MigrationsAssembly(typeof(ApplicationDbContext).Assembly.FullName)));
+            try
+            {
+                // Configure DbContext
+                var connectionString = configuration.GetConnectionString("DefaultConnection")
+                    ?? throw new InvalidOperationException("Connection string 'DefaultConnection' not found.");
+                    
+                services.AddDbContext<ApplicationDbContext>(options =>
+                    options.UseSqlServer(
+                        connectionString,
+                        b => b.MigrationsAssembly(typeof(ApplicationDbContext).Assembly.FullName)));
 
-            // Configure Identity
-            services.AddIdentity<ApplicationUser, ApplicationRole>(options =>
-            {
-                options.Password.RequiredLength = 8;
-                options.Password.RequireDigit = true;
-                options.Password.RequireLowercase = true;
-                options.Password.RequireUppercase = true;
-                options.Password.RequireNonAlphanumeric = true;
-                options.User.RequireUniqueEmail = true;
-                options.Lockout.DefaultLockoutTimeSpan = TimeSpan.FromMinutes(15);
-                options.Lockout.MaxFailedAccessAttempts = 5;
-                options.SignIn.RequireConfirmedEmail = true;
-            })
-            .AddEntityFrameworkStores<ApplicationDbContext>()
-            .AddDefaultTokenProviders();
-
-            // Configure JWT Authentication
-            services.AddAuthentication(options =>
-            {
-                options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-            })
-            .AddJwtBearer(options =>
-            {
-                options.TokenValidationParameters = new TokenValidationParameters
+                // Configure Identity
+                services.AddIdentity<ApplicationUser, ApplicationRole>(options =>
                 {
-                    ValidateIssuer = true,
-                    ValidateAudience = true,
-                    ValidateLifetime = true,
-                    ValidateIssuerSigningKey = true,
-                    ValidIssuer = configuration["JwtSettings:Issuer"],
-                    ValidAudience = configuration["JwtSettings:Audience"],
-                    IssuerSigningKey = new SymmetricSecurityKey(
-                        Encoding.UTF8.GetBytes(configuration["JwtSettings:Key"]))
-                };
-            });
+                    options.Password.RequiredLength = 8;
+                    options.Password.RequireDigit = true;
+                    options.Password.RequireLowercase = true;
+                    options.Password.RequireUppercase = true;
+                    options.Password.RequireNonAlphanumeric = true;
+                    options.User.RequireUniqueEmail = true;
+                    options.Lockout.DefaultLockoutTimeSpan = TimeSpan.FromMinutes(15);
+                    options.Lockout.MaxFailedAccessAttempts = 5;
+                    // Email confirmation disabled for development
+                    options.SignIn.RequireConfirmedEmail = false;
+                })
+                .AddEntityFrameworkStores<ApplicationDbContext>()
+                .AddDefaultTokenProviders();
 
-            // Configure Email Services
-            services.AddSingleton<IEmailTemplateService, EmailTemplateService>();
-            services.AddScoped<IEmailService, SmtpEmailService>();
+                // Configure JWT Authentication
+                var jwtKey = configuration["JwtSettings:Key"] 
+                    ?? throw new InvalidOperationException("JWT Key not found in configuration.");
+                    
+                services.AddAuthentication(options =>
+                {
+                    options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+                    options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+                })
+                .AddJwtBearer(options =>
+                {
+                    options.TokenValidationParameters = new TokenValidationParameters
+                    {
+                        ValidateIssuer = true,
+                        ValidateAudience = true,
+                        ValidateLifetime = true,
+                        ValidateIssuerSigningKey = true,
+                        ValidIssuer = configuration["JwtSettings:Issuer"],
+                        ValidAudience = configuration["JwtSettings:Audience"],
+                        IssuerSigningKey = new SymmetricSecurityKey(
+                            Encoding.UTF8.GetBytes(jwtKey))
+                    };
+                });
 
-            // Configure Identity Services
-            services.AddScoped<IIdentityService, IdentityService>();
+                // Configure Email Services - commented out for development
+                services.AddSingleton<IEmailTemplateService, EmailTemplateService>();
+                services.AddScoped<IEmailService, SmtpEmailService>();
 
-            // Configure Repositories
-            services.AddScoped(typeof(IRepository<>), typeof(Repository<>));
-            services.AddScoped<IUnitOfWork, UnitOfWork>();
+                // Configure Identity Services
+                services.AddScoped<IIdentityService, IdentityService>();
 
-            // Configure Logging
-            var logger = LoggerService.CreateLogger(configuration);
-            services.AddSingleton(logger);
-            services.AddLogging(loggingBuilder => loggingBuilder.AddSerilog(logger, dispose: true));
+                // Configure Repositories
+                services.AddScoped(typeof(IRepository<>), typeof(Repository<>));
+                services.AddScoped<IUnitOfWork, UnitOfWork>();
 
-            return services;
+                // Configure Logging
+                var logger = LoggerService.CreateLogger(configuration);
+                services.AddSingleton(logger);
+                services.AddLogging(loggingBuilder => loggingBuilder.AddSerilog(logger, dispose: true));
+
+                return services;
+            }
+            catch (Exception ex)
+            {
+                // Using Serilog instead of throwing to avoid crashing the application during startup
+                Log.Error(ex, "Error configuring infrastructure services");
+                throw; // Rethrow after logging for proper startup error handling
+            }
         }
     }
 } 
