@@ -4,18 +4,18 @@ namespace PayrollServer.Infrastructure.Services.Email
 {
     public class EmailMessage
     {
-        public string To { get; set; }
-        public string From { get; set; }
-        public string Subject { get; set; }
-        public string Body { get; set; }
+        public required string To { get; set; }
+        public string? From { get; set; }
+        public required string Subject { get; set; }
+        public required string Body { get; set; }
         public bool IsHtml { get; set; } = true;
         public List<EmailAttachment> Attachments { get; set; } = new List<EmailAttachment>();
     }
 
     public class EmailAttachment
     {
-        public string FileName { get; set; }
-        public byte[] Content { get; set; }
-        public string ContentType { get; set; }
+        public required string FileName { get; set; }
+        public required byte[] Content { get; set; }
+        public required string ContentType { get; set; }
     }
 } 
