@@ -5,6 +5,6 @@ namespace PayrollServer.Infrastructure.Services.Email
     public interface IEmailService
     {
         Task SendEmailAsync(EmailMessage emailMessage);
-        Task SendTemplatedEmailAsync(string to, string templateName, object model, string subject = null);
+        Task SendTemplatedEmailAsync(string to, string templateName, object model, string? subject = null);
     }
 } 
